add extension support move simplegui to other extension package
