+++ README.md
@@ -17,7 +17,7 @@ $ pip install schemdule
 
 It's a pure python script, so you can use any python statement in it.
 
-Schemdule provide `at`, `cycle`, and `load` functions for registering events, and a `PrompterConfiger` variable named `prompter` to config prompter (the default prompter is messagebox).
+Schemdule provide `at`, `cycle`, `load` and `ext` functions for registering events, and a `PrompterConfiger` variable named `prompter` to config prompter (the default prompter is Tkinter messagebox).
 
 ```python
 # raw_time can be {hh:mm} or {hh:mm:ss} or a datetime.time object
@@ -36,6 +36,11 @@ def load(source: str) -> None:
     # load from a schema source code
     ...
 
+def ext(name: str) -> None:
+    # use an extension
+    # provided by packages `schemdule-extensions-{extension name}`
+    ...
+
 # the class of the variable `prompter`
 
 class PrompterConfiger:
@@ -48,8 +53,6 @@ class PrompterConfiger:
     def useConsole(self) -> "PrompterConfiger": ...
 
     def useTkinterMessageBox(self) -> "PrompterConfiger": ...
-
-    def useMessageBox(self, auto_close=False) -> "PrompterConfiger": ...
 
 ```
 
@@ -64,6 +67,7 @@ from schemdule.prompters import Prompter, PrompterHub
 at: Callable[[Union[str, time], str, Any], None]
 cycle: Callable[[Union[str, time], Union[str, time], Union[str, time], Union[str, time], str, Any], None]
 load: Callable[[str], None]
+ext: Callable[[str], None]
 prompter: PrompterConfiger
 
 # Schema
@@ -75,6 +79,7 @@ cycle("8:00", "12:00", "00:30:00", "00:10:00", "Working")
 
 prompter.useTkinterMessageBox()
 # use multiple prompter:
+# ext("simplegui") # use simplegui extension (package schemdule-extensions-simplegui)
 # prompter.useBroadcaster().useConsole().useMessageBox(True)
 ```
