+++ src/extensions/simplegui/setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = schemdule-extensions-simplemessagebox
+name = schemdule-extensions-simplegui
 version = 0.0.4
 author = StardustDL
 author_email = stardustdl@163.com
-description = A tiny tool using script as schema to schedule one day and remind you to do something during a day.
+description = A simple GUI extension for schemdule.
 long_description = file: README.md
 long_description_content_type = text/markdown
 url = https://github.com/StardustDL/schemdule
@@ -29,4 +29,3 @@ python_requires = >=3.6
 install_requires = 
     schemdule>=0.0.4
     PySimpleGUI>=4.45.0
-
