+++ src/main/schemdule/prompters/general.py
@@ -3,7 +3,6 @@ from typing import Any
 import click
 
 from . import Prompter, PromptResult
-import PySimpleGUI as sg
 
 
 class TkinterMessageBoxPrompter(Prompter):
@@ -15,18 +14,6 @@ class TkinterMessageBoxPrompter(Prompter):
         top.withdraw()
         tkinter.messagebox.showinfo(f"Attention {message}", payload)
         top.destroy()
-
-        return PromptResult.Resolved
-
-
-class MessageBoxPrompter(Prompter):
-    def __init__(self, auto_close=False) -> None:
-        super().__init__()
-        self.auto_close = auto_close
-
-    def prompt(self, message: str, payload: Any) -> Any:
-        sg.popup_scrolled(str(payload), title=f"Attention {message}", auto_close=self.auto_close,
-                          keep_on_top=True, background_color='white', text_color='black')
 
         return PromptResult.Resolved
