+++ src/main/setup.cfg
@@ -29,7 +29,6 @@ python_requires = >=3.6
 install_requires = 
     click>=8.0.1
     enlighten>=1.10.1
-    PySimpleGUI>=4.45.0
 
 [options.entry_points]
 console_scripts =
